v0.0.7.9 : final push of this version
"Release v0.0.7.9."

+++ src/org/dungeon/io/Sleeper.java
@@ -6,7 +6,7 @@ public final class Sleeper {
 		throw new AssertionError();
 	}
 	
-	public static void sleep(int milliseconds) {
+	public static void sleep(long milliseconds) {
 		if (milliseconds > 0) {
 			try {
 				Thread.sleep(milliseconds);
